+++ backend/src/test/java/ma/iorecycling/service/ReportServiceTest.java
@@ -28,8 +28,8 @@ class ReportServiceTest {
         ValorSummaryDTO summary = new ValorSummaryDTO(
             "2024-01",
             Arrays.asList(
-                new ValorSummaryRowDTO("CARTON", 100.0, new BigDecimal("1.50"), new BigDecimal("150.00")),
-                new ValorSummaryRowDTO("PLASTIQUE", 50.0, new BigDecimal("2.00"), new BigDecimal("100.00"))
+                new ValorSummaryRowDTO("CARTON", BigDecimal.valueOf(100.0), new BigDecimal("1.50"), new BigDecimal("150.00")),
+                new ValorSummaryRowDTO("PLASTIQUE", BigDecimal.valueOf(50.0), new BigDecimal("2.00"), new BigDecimal("100.00"))
             ),
             new BigDecimal("250.00"),
             "MAD"
